requirements.txt
fastapi
uvicorn[standard]  # Needed for WebSocket support
websockets
redis
sqlalchemy
pydantic-settings